Made the requested changes on the old pull-request; This involved deleting some widget/displayable pairs;
