+++ app/src/main/java/cm/aptoide/pt/permission/DialogPermissions.java
@@ -21,7 +21,6 @@ import android.widget.TextView;
 import cm.aptoide.pt.R;
 import cm.aptoide.pt.app.AppViewViewModel;
 import cm.aptoide.pt.dataprovider.model.v7.GetApp;
-import cm.aptoide.pt.dataprovider.model.v7.GetAppMeta;
 import cm.aptoide.pt.permissions.ApkPermission;
 import cm.aptoide.pt.permissions.ApkPermissionGroup;
 import cm.aptoide.pt.util.AppUtils;
@@ -45,30 +44,13 @@ public class DialogPermissions extends DialogFragment {
   private String icon;
   private String size;
 
-  public static DialogPermissions newInstance(GetApp getApp) {
-    GetAppMeta.App app = getApp.getNodes()
-        .getMeta()
-        .getData();
+  public static DialogPermissions newInstance(String appName, String versionName, String icon,
+      String size) {
     DialogPermissions dialog = new DialogPermissions();
-    dialog.getApp = getApp;
-    dialog.appName = app.getName();
-    dialog.versionName = app.getFile()
-        .getVername();
-    dialog.icon = app.getIcon();
-    dialog.size = AptoideUtils.StringU.formatBytes(AppUtils.sumFileSizes(app.getFile()
-        .getFilesize(), app.getObb()), false);
-    return dialog;
-  }
-
-  public static DialogPermissions newInstance(AppViewViewModel app) {
-    DialogPermissions dialog = new DialogPermissions();
-    dialog.app = app;
-    dialog.appName = app.getAppName();
-    dialog.versionName = app.getVersionName();
-    dialog.icon = app.getIcon();
-    dialog.size =
-        AptoideUtils.StringU.formatBytes(AppUtils.sumFileSizes(app.getFileSize(), app.getObb()),
-            false);
+    dialog.appName = appName;
+    dialog.versionName = versionName;
+    dialog.icon = icon;
+    dialog.size = size;
     return dialog;
   }
