+++ app/src/main/res/layout/appview_app_description_layout.xml
@@ -10,8 +10,7 @@
       android:layout_width="wrap_content"
       android:layout_height="wrap_content"
       android:text="@string/description"
-      android:textColor="@color/black"
-      android:textSize="16sp"
+      style="@style/Aptoide.TextView.Regular.M.BlackAlpha"
       />
 
   <TextView
@@ -21,8 +20,8 @@
       android:layout_marginTop="8dp"
       android:ellipsize="end"
       android:maxLines="3"
-      android:textSize="12sp"
       tools:text="As a data-driven company, we rely heavily on experiments to guide products and features. At any given time, we have around 1,000 experiments running, and we’re adding more every day. Because we’re constantly increasing the number of experiments and logging corresponding data, we need a reliable, simple to use platform engineers can use without error"
+      style="@style/Aptoide.TextView.Regular.XS"
       />
 
   <Button
