+++ app/src/main/res/layout/appview_uploaded_by_layout.xml
@@ -10,8 +10,7 @@
       android:layout_width="wrap_content"
       android:layout_height="wrap_content"
       android:text="@string/appview_uploaded_by"
-      android:textColor="@color/black"
-      android:textSize="16sp"
+      style="@style/Aptoide.TextView.Regular.M.BlackAlpha"
       />
 
   <include
@@ -52,9 +51,8 @@
           android:layout_height="wrap_content"
           android:ellipsize="end"
           android:maxLines="1"
-          android:textColor="@color/black"
-          android:textSize="16sp"
           tools:text="Dead Pirates store"
+          style="@style/Aptoide.TextView.Regular.M.BlackAlpha"
           />
 
       <ImageView
@@ -64,7 +62,7 @@
           android:layout_below="@id/store_name"
           android:layout_marginTop="5dp"
           android:src="@drawable/ic_user_shape"
-          android:tint="@color/grey_700"
+          android:tint="@color/grey_medium"
           />
 
       <TextView
@@ -77,9 +75,8 @@
           android:layout_marginTop="3dp"
           android:layout_toEndOf="@id/user_shape"
           android:layout_toRightOf="@id/user_shape"
-          android:textColor="@color/grey_700"
-          android:textSize="10sp"
           tools:text="348 K"
+          style="@style/Aptoide.TextView.Regular.XXS.GreyMedium"
           />
 
       <ImageView
@@ -93,7 +90,7 @@
           android:layout_toEndOf="@id/user_count"
           android:layout_toRightOf="@id/user_count"
           android:src="@drawable/ic_action_download"
-          android:tint="@color/grey_700"
+          android:tint="@color/grey_medium"
           />
 
       <TextView
@@ -106,9 +103,8 @@
           android:layout_marginTop="3dp"
           android:layout_toEndOf="@id/downloads_shape"
           android:layout_toRightOf="@id/downloads_shape"
-          android:textColor="@color/grey_700"
-          android:textSize="10sp"
           tools:text="348 K"
+          style="@style/Aptoide.TextView.Regular.XXS.GreyMedium"
           />
     </RelativeLayout>
