+++ app/src/main/res/layout/displayable_app_view_developer.xml
@@ -25,8 +25,7 @@
       android:paddingRight="17.3dp"
       android:paddingStart="13.3dp"
       android:text="@string/additional_information"
-      android:textColor="@color/black"
-      android:textSize="16sp"
+      style="@style/Aptoide.TextView.Regular.M.BlackAlpha"
       />
 
   <LinearLayout
